requirements.txt
whisper
pydub
torch
tqdm
python-docx
requests